allowed non-erasable syntax
