+++ src/types/ProjectLink.ts
@@ -1,0 +1,7 @@
+import type { ProjectLinkVariant } from "./ProjectLinkVarient";
+
+export interface ProjectLink {
+  name: string;
+  url: string;
+  variant: ProjectLinkVariant;
+}
