+++ src/types/ProjectLinkVarient.ts
@@ -1,0 +1,4 @@
+export enum ProjectLinkVariant {
+  Neutral = "neutral",
+  Default = "default",
+}
